feat: Deprecate change/slide in favor of on-change/on-slide
Also renamed "set" action to "on-set" without deprecating "set".

If you didn't provide a "set" action, Ember would try and use the
standard "set" function of the component as the action which would blow
up.

This poor naming convention lead to the deprecation of the old events
across the board in favor of "on-*" names.